moving createElement to util file
